backend/requirements.txt
streamlit>=1.28.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
huggingface-hub>=0.19.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
youtube-transcript-api>=0.6.1
